refactor: change modules folder to components folder
